refactor: add views to input and output view object files
